tools/marketplace-automation/requirements.txt
Pillow>=10.0.0
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.0.0
openai>=1.0.0
rembg>=2.0.50
onnxruntime>=1.16.0
tqdm>=4.66.0
